feat: Implement initial dashboard structure including command center page, create project modal, and toggle component.
